Test `StoryItemMapper` in isolation
